+++ muna/beta/openai/__init__.py
@@ -1,6 +1,6 @@
 # 
 #   Muna
-#   Copyright © 2025 NatML Inc. All Rights Reserved.
+#   Copyright © 2026 NatML Inc. All Rights Reserved.
 #
 
 from .annotations import Annotations
@@ -8,5 +8,6 @@ from .openai import OpenAIClient
 from .schema import (
     ChatCompletion, ChatCompletionChunk, ChatCompletionReasoningEffort,
     Choice, DeltaMessage, EmbeddingCreateResponse, Embedding, Message,
-    SpeechCreateResponse, StreamChoice
+    SpeechCreateResponse, SpeechResponseFormat, SpeechStreamFormat,
+    StreamChoice
 )